Restructure and refactor for windows builds
