+++ backend/support/qrcode/libqrencode/.gitignore
