+++ backend/support/qrcode/libqrencode/.travis.yml
