+++ backend/support/qrcode/libqrencode/COPYING
