+++ backend/support/qrcode/libqrencode/Doxyfile
