+++ backend/support/qrcode/libqrencode/acinclude.m4
