+++ backend/support/qrcode/libqrencode/autogen.sh
