+++ backend/support/qrcode/libqrencode/bitstream.c
