+++ backend/support/qrcode/libqrencode/bitstream.h
