+++ backend/support/qrcode/libqrencode/configure.ac
