+++ backend/support/qrcode/libqrencode/libqrencode.pc.in
