+++ backend/support/qrcode/libqrencode/mask.c
