+++ backend/support/qrcode/libqrencode/mqrspec.c
