+++ backend/support/qrcode/libqrencode/mqrspec.h
