+++ backend/support/qrcode/libqrencode/qrenc.c
