+++ backend/support/qrcode/libqrencode/qrencode.1.in
