+++ backend/support/qrcode/libqrencode/qrencode.c
