+++ backend/support/qrcode/libqrencode/qrencode.h
