+++ backend/support/qrcode/libqrencode/qrencode.spec.in
