+++ backend/support/qrcode/libqrencode/qrencode_inner.h
