+++ backend/support/qrcode/libqrencode/qrinput.c
