+++ backend/support/qrcode/libqrencode/qrinput.h
