+++ backend/support/qrcode/libqrencode/qrspec.c
