+++ backend/support/qrcode/libqrencode/qrspec.h
