+++ backend/support/qrcode/libqrencode/rsecc.c
