+++ backend/support/qrcode/libqrencode/rsecc.h
