+++ backend/support/qrcode/libqrencode/split.c
