+++ backend/support/qrcode/libqrencode/split.h
