+++ backend/support/qrcode/libqrencode/use/config.rpath
