--- backend/support/qrcode/qt-qrcode/LICENSE_HEADER.txt
@@ -1,26 +1,0 @@
-/****************************************************************************
- *
- * Copyright (c) 2015 Daniel San, All rights reserved.
- * 
- *    Contact: daniel.samrocha@gmail.com
- *       File: %FILENAME%
- *     Author: %USER%
- * Created on: %DATE%
- *    Version: 
- *
- * This file is part of the Qt QRCode library.
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 3.0 of the License, or (at your option) any later version.
- * 
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details.
- * 
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library.
- *
- ***************************************************************************/
