+++ backend/support/qrcode/qt-qrcode/QtQrCodeQuickItem.cpp
