+++ backend/support/qrcode/qt-qrcode/QtQrCodeQuickItem.hpp
