--- backend/support/qrcode/qt-qrcode/defaults.pri
@@ -1,2 +1,0 @@
-INCLUDEPATH += $$PWD/lib
-SRC_DIR = $$PWD
